Add extra item fields (#153)
* Add more fields on a selected item

- add ability to return an item

Add extra fields for few required fields

- add min and max for weight and height

Add more fields in testdata.sql to associate with a selected event

Add minor backend changes to support more inv fields

- added sample data in testdata.sql
- update existing sql to support more fields
- added ability to filter unwanted columns in the inventory list
- added similar design as shared item component for consistency and bring light on a seperate edit mode for select item.
- added easy edit for toggle selection on returnable field, others stay same. all new fields are editable.
- updated constants to support extra fields for header purposes.
- added more fields to inventory struct
- added the fields in each of the response.

we should probably refactor the sql since we are using the same thing a couple of times. it would be nice to update it only at one location. some of the api fields do not return the added extra fields and finding that and adding that is getting difficult. we should consider to consolidate if possible

- display clear heading components for selected inventories and reduce clutter

* Update prod docker file

- add support to deploy application in prod
- remove console statements
- run migration in prod
- fix empty event list
* update mashed schema docx file

+++ mainDeploy.sh
@@ -25,9 +25,14 @@ loadProdEnv() {
 
     echo "loadProdEnv flag provided. building all containers in production mode."
 
+    chmod +x setup/_loadProdEnvVariables.sh
+    ./setup/_loadProdEnvVariables.sh
+
     docker-compose down
     docker-compose -f docker-compose.deploy.yml up --build -d
 
+    sleep +2
+    loadMigration
 }
 
 loadMigration() {
@@ -40,8 +45,16 @@ loadMigration() {
 
 uninstall() {
 
-    echo "uninstalling software. removing all data..."
-    docker-compose down --remove-orphans --volumes
+    echo " WARNING !!!! PRODUCTION INSTANCE:: PURGING ALL DATA ... Have you backed up? Press 'Y' to confirm, or 'N' to cancel."
+    read -r -p "Are you sure you want to proceed? [Y/n] " response
+    response=$(echo "$response" | tr '[:upper:]' '[:lower:]') 
+    if [[ "$response" =~ ^(yes|y| ) ]]; then
+        echo "PRODUCTION INSTANCE:: PURGING ALL DATA ... hope you know what you are doing..."
+        sleep 10
+        docker-compose down --remove-orphans --volumes
+    else
+        echo "Uninstallation canceled."
+    fi
 }
 
 if [ "$#" -eq 0 ]; then
